adding twirl + some improvements
