+++ project/plugins.sbt
@@ -5,3 +5,5 @@ addSbtPlugin("io.spray" % "sbt-revolver" % "0.7.2")
 addSbtPlugin("com.typesafe.sbt" % "sbt-native-packager" % "1.0.0")
 
 addSbtPlugin("org.scala-js" % "sbt-scalajs" % "0.6.3")
+
+addSbtPlugin("com.typesafe.sbt" % "sbt-twirl" % "1.1.1")
